chore: `extending-the-sdk` project for resources
