+++ docs/resources/extending-the-sdk/MyResourceDetector.cs
@@ -1,0 +1,17 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using OpenTelemetry.Resources;
+
+internal class MyResourceDetector : IResourceDetector
+{
+    public Resource Detect()
+    {
+        var attributes = new List<KeyValuePair<string, object>>
+        {
+            new KeyValuePair<string, object>("key", "val"),
+        };
+
+        return new Resource(attributes);
+    }
+}
