+++ docs/resources/extending-the-sdk/Program.cs
@@ -1,41 +1,6 @@
-# Resources
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
 
-## Resource Detector
-
-OpenTelemetry .NET SDK provides a resource detector for detecting resource
-information from the `OTEL_RESOURCE_ATTRIBUTES` and `OTEL_SERVICE_NAME`
-environment variables.
-
-Custom resource detectors can be implemented:
-
-* ResourceDetectors should inherit from
-  `OpenTelemetry.Resources.IResourceDetector`, (which belongs to the
-  [OpenTelemetry](../../src/OpenTelemetry/README.md) package), and implement
-  the `Detect` method.
-
-A demo `ResourceDetector` is shown [here](./extending-the-sdk/MyResourceDetector.cs):
-
-```csharp
-using OpenTelemetry.Resources;
-
-internal class MyResourceDetector : IResourceDetector
-{
-    public Resource Detect()
-    {
-        var attributes = new List<KeyValuePair<string, object>>
-        {
-            new KeyValuePair<string, object>("key", "val"),
-        };
-
-        return new Resource(attributes);
-    }
-}
-```
-
-There are two different ways to add the custom `ResourceDetector` to the
-OTEL signals, via the `Sdk.Create` approach:
-
-```csharp
 using System.Diagnostics;
 using System.Diagnostics.Metrics;
 using Microsoft.Extensions.Logging;
@@ -93,17 +58,3 @@ public class Program
             .LogInformation("Hello from {name} {price}.", "tomato", 2.99);
     }
 }
-```
-
-or via `OpenTelemetry.Extensions.Hosting` method:
-
-```csharp
-    services.AddSingleton<MyResourceDetector>();
-
-    services.AddOpenTelemetry()
-        .ConfigureResource(builder => builder
-            .AddDetector(sp =>
-                sp.GetRequiredService<MyResourceDetector>()))
-        .WithTracing(builder => builder.AddConsoleExporter())
-        .WithMetrics(builder => builder.AddConsoleExporter());
-```
